\name{RXapian-package}
\alias{RXapian-package}
\alias{RXapian}
\docType{package}
\title{
\packageTitle{RXapian}
}
\description{
\packageDescription{RXapian}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RXapian}
\packageIndices{RXapian}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RXapian}

Maintainer: \packageMaintainer{RXapian}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
