% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_dbAvlength}
\alias{xapian_dbAvlength}
\title{Database avlength}
\usage{
xapian_dbAvlength(dbpath)
}
\arguments{
\item{dbpath}{path to a Xapian database}
}
\value{
average length of the documents in the database
}
\description{
Database avlength
}
\examples{
\dontrun{
db<- c("path/to/database")
xapian_dbAvlength(db)
}
}

