% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_dbDoccount}
\alias{xapian_dbDoccount}
\title{Database doccount}
\usage{
xapian_dbDoccount(dbpath)
}
\arguments{
\item{dbpath}{path to a Xapian database}
}
\value{
number of documents in the database
}
\description{
Database doccount
}
\examples{
\dontrun{
db<- c("path/to/database")
xapian_dbDoccount(db)
}
}

