% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_dbDoclength_upper_bound}
\alias{xapian_dbDoclength_upper_bound}
\title{Database doclength_upper_bound}
\usage{
xapian_dbDoclength_upper_bound(dbpath)
}
\arguments{
\item{dbpath}{path to a Xapian database}
}
\value{
an upper bound on the length of a document in the database
}
\description{
Database doclength_upper_bound
}
\examples{
\dontrun{
db<- c("path/to/database")
xapian_dbDoclength_upper_bound(db)
}
}

