% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_dbHas_positions}
\alias{xapian_dbHas_positions}
\title{Database has_positions}
\usage{
xapian_dbHas_positions(dbpath)
}
\arguments{
\item{dbpath}{path to a Xapian database}
}
\value{
whether this database has any positional information
}
\description{
Database has_positions
}
\examples{
\dontrun{
db<- c("path/to/database")
xapian_dbHas_positions(db)
}
}

