% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_dbInfo}
\alias{xapian_dbInfo}
\title{DatabaseInfo}
\usage{
xapian_dbInfo(dbpath)
}
\arguments{
\item{dbpath}{path to a Xapian database}
}
\value{
list of information about the Xapian database at specified location
}
\description{
DatabaseInfo
}
\examples{
\dontrun{
db<- c("path/to/database")
xapian_dbInfo(db)
}
}

