% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_dbLastdocid}
\alias{xapian_dbLastdocid}
\title{Database lastdocid}
\usage{
xapian_dbLastdocid(dbpath)
}
\arguments{
\item{dbpath}{path to a Xapian database}
}
\value{
highest document id which has been used in the database
}
\description{
Database lastdocid
}
\examples{
\dontrun{
db<- c("path/to/database")
xapian_dbLastdocid(db)
}
}

