% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_dbMetadata}
\alias{xapian_dbMetadata}
\title{Database Metadata}
\usage{
xapian_dbMetadata(dbpath, key)
}
\arguments{
\item{dbpath}{path to a Xapian database}

\item{key}{key of the metadata item to access}
}
\value{
user-specified metadata associated with the given key
}
\description{
Database Metadata
}
\examples{
\dontrun{
db<- c("path/to/database")
key<-c("k1")
xapian_dbMetadata(db,key)
}
}

