% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_dbsetMetadata}
\alias{xapian_dbsetMetadata}
\title{Database setmetadata}
\usage{
xapian_dbsetMetadata(dbpath, key, value)
}
\arguments{
\item{dbpath}{path to a Xapian database}

\item{key}{key of the metadata item to set}

\item{value}{value of the metadata item to set}
}
\value{
none
}
\description{
Database setmetadata
}
\examples{
\dontrun{
db<- c("path/to/database")
key<-c("k1")
value<-c("v1")
xapian_dbsetMetadata(db,key,value)
}
}

