% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_dbSpelling_suggestion}
\alias{xapian_dbSpelling_suggestion}
\title{Database spelling_suggestion}
\usage{
xapian_dbSpelling_suggestion(dbpath, arg)
}
\arguments{
\item{dbpath}{path to a Xapian database}

\item{arg}{potentially misspelled word}
}
\value{
a spelling correction to a potentially misspelled word
}
\description{
Database spelling_suggestion
}
\examples{
\dontrun{
db<- c("path/to/database")
arg<-list(word="helllo", max_edit_distance = 3)
xapian_dbSpelling_suggestion(db,arg)
}
}

