% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_dbUuid}
\alias{xapian_dbUuid}
\title{Database uuid}
\usage{
xapian_dbUuid(dbpath)
}
\arguments{
\item{dbpath}{path to a Xapian database}
}
\value{
a UUID for the database
}
\description{
Database uuid
}
\examples{
\dontrun{
db<- c("path/to/database")
xapian_dbUuid(db)
}
}

