% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_dbValue_freq}
\alias{xapian_dbValue_freq}
\title{Database value_freq}
\usage{
xapian_dbValue_freq(dbpath, slot)
}
\arguments{
\item{dbpath}{path to a Xapian database}

\item{slot}{value slot to examine}
}
\value{
frequency of the given value slot
}
\description{
Database value_freq
}
\examples{
\dontrun{
db<- c("path/to/database")
slot<-c(1)
xapian_dbValue_freq(db,slot)
}
}

