% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_dbValue_upper_bound}
\alias{xapian_dbValue_upper_bound}
\title{Database value_upper_bound}
\usage{
xapian_dbValue_upper_bound(dbpath, slot)
}
\arguments{
\item{dbpath}{path to a Xapian database}

\item{slot}{value slot to examine}
}
\value{
an upper bound on the values stored in the given value slot
}
\description{
Database value_upper_bound
}
\examples{
\dontrun{
db<- c("path/to/database")
slot<-c(1)
xapian_dbValue_upper_bound(db,slot)
}
}

