% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_delete_by_term}
\alias{xapian_delete_by_term}
\title{xapian_delete_by_term}
\usage{
xapian_delete_by_term(dbpath, unique_term)
}
\arguments{
\item{dbpath}{path to a Xapian database}

\item{unique_term}{The term to remove references to}
}
\value{
none
}
\description{
xapian_delete_by_term
}
\examples{
\dontrun{
db<- c("path/to/database")
term <- "hello"
xapian_delete_by_term(dbpath,term)
}
}

