% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexingWrapper.R
\name{xapian_index}
\alias{xapian_index}
\title{indexing wrapper}
\usage{
xapian_index(dbpath, dataFrame, idColumn = NULL, indexFields = NULL,
  filterFields = NULL, stemmer = c("english", "en", "danish", "da", "dutch",
  "nl", "english_lovins", "lovins", "english_porter", "porter", "finnish", "fi",
  "french", "fr", "german", "de", "german2", "hungarian", "hu", "italian", "it",
  "kraaij_pohlmann", "norwegian", "nb", "nn", "no", "portuguese", "pt",
  "romanian", "ro", "russian", "ru", "spanish", "es", "swedish", "sv",
  "turkish", "tr"))
}
\arguments{
\item{dbpath}{path to a Xapian database}

\item{dataFrame}{data frame to be indexed}

\item{idColumn}{index of a column in the data frame whose row value will be used as a unique identifier to each document(row)}

\item{indexFields}{list of 'list of fields' that will be indexed using a Xapian::TermGenerator}

\item{filterFields}{list with information on boolean terms to be added to Xapian::Document}

\item{stemmer}{the stemmer that should be applied to the Xapian::TermGenerator}
}
\value{
none
}
\description{
indexing wrapper
}
\examples{
\dontrun{

db<- c("path/to/database")
data<-read.csv("path/to/csv") or data<-data.frame(..)
id<-c(0)
indexFields<-list(list(index=2,prefix="S"), list(index=8,prefix="XD"))
filterFields<-list(list(index=6, prefix="XM", separator=";"))
xapian_index(db,data,id,indexFields,filterFields,"en")

}
}

