% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchWrapper.R
\name{xapian_search}
\alias{xapian_search}
\title{search wrapper}
\usage{
xapian_search(dbpath = NULL, enquireList = NULL, queryList = NULL)
}
\arguments{
\item{dbpath}{path to a Xapian database}

\item{enquireList}{list of arguments for Xapian::Enquire}

\item{queryList}{a list}
}
\value{
a data frame or a list of lists containing search results
}
\description{
search wrapper
}
\examples{
\dontrun{

db<- c("path/to/database")

query<-list( queryString="watch", 
             prefix.fields=list( list(prefix="S",name="title"), 
                                 list(prefix="XD",name="description")), 
             stemmer="en")

enq<-list(matchspy=c(0,1))

res<-xapian_search(db,enq,query)

}
}

