\name{RXapian-package}
\alias{RXapian-package}
\alias{RXapian}
\docType{package}
\title{
 R interface to Xapian search engine library.
}
\description{
 This package provides an interface to the advanced indexing and retrieval operations of Xapian search engine library.
}
\references{Xapian search engine library: \url{https://xapian.org/}}
