% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseWrapper.R
\name{xapian_delete}
\alias{xapian_delete}
\title{xapian_delete}
\usage{
xapian_delete(dbpath, docid = NULL, unique_term = NULL)
}
\arguments{
\item{dbpath}{path to a Xapian database}

\item{docid}{the document ID of the document to be removed}

\item{unique_term}{The term to remove references to}
}
\value{
none
}
\description{
xapian_delete
}
\examples{
\dontrun{
db<- c("path/to/database")
docid <- 20
xapian_delete(dbpath= db,docid= docid)
}
}

